+++ authentication.php
@@ -4,7 +4,13 @@ if ($_SESSION['authid'] == "marksman") {
 	$username = $_SESSION['username'];
 } else {
     if ($force_login_redirect == true) {
-        header("Location: ./login.php");
+        if (file_exists("./login.php")) {
+            header("Location: ./login.php");
+        } else if (file_exists("../login.php")) {
+            header("Location: ../login.php");
+        } else {
+            header("Location: ./login.php");
+        }
         exit();
     }
 }