+++ utils.php
@@ -1,5 +1,11 @@
 <?php
-include "./config.php";
+if (file_exists("./config.php")) {
+    include "./config.php";
+} else if (file_exists("../config.php")) {
+    include "../config.php";
+} else if (file_exists("../../config.php")) {
+    include "../../config.php";
+}